+++ omg/test/bit_helper.cpp
@@ -15,5 +15,6 @@ QUIZ_CASE(omg_bit_helper) {
   quiz_assert(BitHelper::countLeadingZeros(0x00F00000) == 8);
   quiz_assert(BitHelper::numberOfOnes(0x00F00000) == 4);
   quiz_assert(BitHelper::indexOfMostSignificantBit(0x00F00000) == 23);
-  quiz_assert(BitHelper::numberOfBitsToCountUpTo(16) == 4);
+  quiz_assert(BitHelper::numberOfBitsToCountUpTo(static_cast<uint32_t>(16)) ==
+              4);
 }
